pyproject.toml
[project]
name = "monitor-the-situation-stellar"
version = "0.1.0"
description = "AI agent-powered geopolitical market intelligence, paid via x402 on Stellar"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "hyperliquid-python-sdk>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.28.0",
]
